// src/app/models/automotive-part-vehicle.ts
export class AutomotivePartVehicle{
    AutomotivePartVehicleID:number;
    AutomotivePartID:number;
    EnglishModelName:string;
    EnglishMakeName:string;
    ArabicMakeName: string;
    ArabicModelName: string;
    MakeID:number;
    ModelID:number;
    ProductionYear:number;
    EngineNumber: string;
    EngineSize:number;
    EngineSizeUnit :string;
    RequiredQuantity:number;
    SeriesYearStart:number;
    SeriesYearEnd:number;
    ModelVersionID:number;
    IsModified: boolean;
    EngineTypeID: number;
    VehicleTypeID: number;
    VehicleTypeName: string;
    EngineTypeName:string;
    FaceLift: boolean;
    FaceLiftTypeID: number;
    EngineCapacity: number;
    VehicleID: number;
    GroupName: string;
    EngineCode : string;
    SeriesID : string;
    MinYear: number;
    MaxYear : number;
}